fixed the adding problem
